# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dondok API
  version: 0.10.0
  description: 돈독 프론트엔드와 백엔드 사이의 HTTP 계약
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/ledger-books/current:
    get:
      summary: 현재 사용자의 가계부와 구성원 조회
      operationId: getCurrentLedgerBook
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 가계부가 없으면 ledger는 null
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentLedgerBook" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      summary: 현재 가계부 전체 삭제
      description: 모든 구성원이 같은 권한으로 실행할 수 있다. 가계부 하위 데이터와 멤버십은 한 트랜잭션에서 물리 삭제하지만 사용자 계정과 로그인 세션은 유지한다.
      operationId: deleteCurrentLedgerBook
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LedgerBookDeletionRequest" }
      responses:
        "204": { description: 가계부와 모든 하위 데이터 삭제 완료 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/ledger-books:
    post:
      summary: 현재 계정의 가계부와 생성자 멤버십 생성
      operationId: createLedgerBook
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "201":
          description: 생성된 가계부와 생성자 멤버십
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerBook" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/ledger-books/current/invitations:
    get:
      summary: 현재 가계부의 발급 초대 목록
      operationId: getLedgerInvitations
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 원문 token을 제외한 초대 상태 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerInvitationSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: 7일·1회용 초대 URL과 코드 발급
      operationId: createLedgerInvitation
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "201":
          description: 원문 code와 URL은 이 응답에서만 제공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssuedLedgerInvitation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/ledger-books/current/invitations/{invitationId}:
    delete:
      summary: 사용 가능한 초대 취소
      operationId: revokeLedgerInvitation
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - name: invitationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 초대 취소 완료 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/ledger-invitations/preview:
    post:
      summary: 초대 URL 또는 직접 코드 미리보기
      operationId: previewLedgerInvitation
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCodeRequest" }
      responses:
        "200":
          description: token을 소비하지 않은 참여 확인 정보
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerInvitationPreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/ledger-invitations/redemptions:
    post:
      summary: 초대 수락과 즉시 가계부 참여
      operationId: redeemLedgerInvitation
      tags: [Membership]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCodeRequest" }
      responses:
        "201":
          description: 참여한 가계부와 갱신된 구성원
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerBook" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/assets/{cardAssetId}/card-statements:
    parameters:
      - name: cardAssetId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 카드 자산의 미결제 중심 명세 cursor 목록
      description: 같은 가계부의 보관된 카드도 과거 명세 조회를 허용한다. 기본값은 OPEN·FINALIZED만 반환하며 includePaid=true이면 최근 PAID도 포함한다.
      operationId: getCardStatements
      tags: [CardStatement]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: includePaid
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: 결제일·식별자 내림차순의 bounded 명세 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardStatementPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/card-statements/{statementId}:
    parameters:
      - name: statementId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 카드 명세와 결제 내역 authoritative 상세
      operationId: getCardStatement
      tags: [CardStatement]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 현재 미결제액과 카드에 설정된 결제 계좌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardStatementDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/card-statements/{statementId}/prepayments:
    parameters:
      - name: statementId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 결제일 전 카드 명세 부분·복수 선결제
      description: 카드에 현재 설정된 결제 계좌에서 카드로 자산만 이동하며 수입·지출 통계에는 포함하지 않는다.
      operationId: createCardStatementPrepayment
      tags: [CardStatement]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyCardStatementPrepaymentRequest" }
      responses:
        "201":
          description: 선결제 후 명세·결제·이체 거래 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardStatementPaymentResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/card-statements/{statementId}/prepayments/preview:
    parameters:
      - name: statementId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 카드 명세 선결제 후 잔액 서버 미리보기
      description: 계좌 잔액은 정상적인 다른 거래와 공존하므로 preview token의 stale 기준에서 제외하고 apply 시 최신 잔액으로 다시 계산한다.
      operationId: previewCardStatementPrepayment
      tags: [CardStatement]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCardStatementPrepaymentRequest" }
      responses:
        "200":
          description: 명세·적용일·결제 계좌 기준 authoritative 미리보기
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardStatementPrepaymentPreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: DONDOK_SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 같은 사용자의 동일 키·동일 payload 재시도는 기존 리소스의 동일 명령 결과를 반환한다.
      schema: { type: string, minLength: 1, maxLength: 100 }
  schemas:
...
    CurrentLedgerBook:
      type: object
      required: [ledger]
      properties:
        ledger:
          oneOf:
            - { $ref: "#/components/schemas/LedgerBook" }
            - { type: "null" }
    LedgerBook:
      type: object
      required: [ledgerId, version, members]
      properties:
        ledgerId: { type: string, format: uuid }
        version: { type: integer, format: int64 }
        members:
          type: array
          items: { $ref: "#/components/schemas/LedgerMember" }
    LedgerBookDeletionRequest:
      type: object
      required: [expectedLedgerId, expectedVersion, confirmationPhrase]
      properties:
        expectedLedgerId:
          type: string
          format: uuid
          description: 삭제 확인 dialog를 열 때 보던 가계부 식별자. 삭제 후 새 가계부가 생긴 ABA 경합에서 새 가계부 삭제를 방지한다.
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        confirmationPhrase:
          type: string
          const: 가계부 삭제
          description: 실수로 전체 데이터를 지우지 않도록 사용자가 직접 입력한 확인 문구
    LedgerMember:
      type: object
      required: [memberId, displayName, joinedAt, currentUser]
      properties:
        memberId: { type: string, format: uuid }
        displayName: { type: string }
        joinedAt: { type: string, format: date-time }
        currentUser: { type: boolean }
    LedgerInvitationSummary:
      type: object
      required: [invitationId, status, createdAt, expiresAt]
      properties:
        invitationId: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, REDEEMED, REVOKED, EXPIRED] }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    IssuedLedgerInvitation:
      allOf:
        - { $ref: "#/components/schemas/LedgerInvitationSummary" }
        - type: object
          required: [code, inviteUrl]
          properties:
            code: { type: string, minLength: 32, maxLength: 100 }
            inviteUrl: { type: string, format: uri }
    InvitationCodeRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 32, maxLength: 100 }
    LedgerInvitationPreview:
      type: object
      required: [memberNames, memberCount, expiresAt]
      properties:
        memberNames:
          type: array
          items: { type: string }
        memberCount: { type: integer, minimum: 1 }
        expiresAt: { type: string, format: date-time }
...
    TransactionType:
      type: string
      enum: [INCOME, EXPENSE, TRANSFER]
    TransferSubtype:
      type: string
      enum: [NORMAL, CARD_SETTLEMENT, CARD_PREPAYMENT]
...
    TransactionCategory:
      type: object
      required: [categoryId, name]
      properties:
        categoryId: { type: string, format: uuid }
        name: { type: string }
    TransactionMember:
      type: object
      required: [memberId, displayName]
      properties:
        memberId: { type: string, format: uuid }
        displayName: { type: string }
    TransactionAsset:
      type: object
      required: [assetId, name]
      properties:
        assetId: { type: string, format: uuid }
        name: { type: string }
    TransactionPosting:
      type: object
      required: [assetId, assetName, deltaWon]
      properties:
        assetId: { type: string, format: uuid }
        assetName: { type: string }
        deltaWon: { type: integer, format: int64 }
    Transaction:
      type: object
      required:
        - transactionId
        - type
        - transferSubtype
        - managementType
        - relatedPurchaseTransactionId
        - occurredOn
        - amountWon
        - category
        - performedBy
        - createdBy
        - asset
        - description
        - postings
        - installmentCount
        - version
        - createdAt
        - updatedAt
      properties:
        transactionId: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/TransactionType" }
        transferSubtype:
          oneOf:
            - { $ref: "#/components/schemas/TransferSubtype" }
            - { type: "null" }
        managementType:
          type: string
          enum: [GENERAL, CARD_PURCHASE, CARD_REFUND, SYSTEM]
        relatedPurchaseTransactionId:
          description: 카드 환불 거래가 연결된 원 구매 식별자이며 그 외에는 null이다.
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
        occurredOn: { type: string, format: date }
        amountWon: { type: integer, format: int64 }
        category:
          oneOf:
            - { $ref: "#/components/schemas/TransactionCategory" }
            - { type: "null" }
        performedBy:
          oneOf:
            - { $ref: "#/components/schemas/TransactionMember" }
            - { type: "null" }
        createdBy:
          oneOf:
            - { $ref: "#/components/schemas/TransactionMember" }
            - { type: "null" }
        asset:
          description: 수입·지출에서 사용자가 선택한 자산이다. 체크카드는 실제 posting 계좌와 다를 수 있다.
          oneOf:
            - { $ref: "#/components/schemas/TransactionAsset" }
            - { type: "null" }
        description:
          oneOf:
            - { type: string }
            - { type: "null" }
        postings:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/TransactionPosting" }
        installmentCount:
          oneOf:
            - { type: integer, minimum: 1 }
            - { type: "null" }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CardStatementPayment:
      type: object
      required: [paymentId, paymentType, settlementAssetId, settlementAssetName, amountWon, returnedAmountWon, effectiveAmountWon, paidOn, settlementTransactionId, createdByMemberId]
      properties:
        paymentId: { type: string, format: uuid }
        paymentType: { type: string, enum: [PREPAYMENT, REGULAR] }
        settlementAssetId: { type: string, format: uuid }
        settlementAssetName: { type: string }
        amountWon: { type: integer, format: int64, minimum: 1 }
        returnedAmountWon: { type: integer, format: int64, minimum: 0 }
        effectiveAmountWon: { type: integer, format: int64, minimum: 0 }
        paidOn: { type: string, format: date }
        settlementTransactionId: { type: string, format: uuid }
        createdByMemberId:
          description: 수동 선결제 작성자이며 자동 정기 결제는 null이다.
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
    CardStatementSummary:
      type: object
      required: [statementId, cardAsset, dueOn, status, grossAmountWon, paidAmountWon, remainingAmountWon, version, automaticSettlement]
      properties:
        statementId: { type: string, format: uuid }
        cardAsset: { $ref: "#/components/schemas/TransactionAsset" }
        dueOn: { type: string, format: date }
        status: { type: string, enum: [OPEN, FINALIZED, PAID, CANCELLED] }
        grossAmountWon: { type: integer, format: int64, minimum: 0 }
        paidAmountWon: { type: integer, format: int64, minimum: 0 }
        remainingAmountWon: { type: integer, format: int64, minimum: 0 }
        version: { type: integer, format: int64, minimum: 0 }
        automaticSettlement:
          oneOf:
            - { $ref: "#/components/schemas/CardPaymentSchedule" }
            - { type: "null" }
    CardStatementDetail:
      allOf:
        - { $ref: "#/components/schemas/CardStatementSummary" }
        - type: object
          required: [prepayableAmountWon, settlementAsset, autoSettlementEnabled, payments]
          properties:
            prepayableAmountWon:
              type: integer
              format: int64
              minimum: 0
              description: 서버 현재 날짜와 명세 상태를 반영해 지금 선결제할 수 있는 최대 금액
            settlementAsset:
              description: 카드에 현재 설정된 단일 결제 계좌이며 미설정이면 null이다.
              oneOf:
                - { $ref: "#/components/schemas/CardSettlementAsset" }
                - { type: "null" }
            autoSettlementEnabled: { type: boolean }
            payments:
              type: array
              items: { $ref: "#/components/schemas/CardStatementPayment" }
    CardSettlementAsset:
      type: object
      required: [assetId, name, currentBalanceWon]
      properties:
        assetId: { type: string, format: uuid }
        name: { type: string }
        currentBalanceWon: { type: integer, format: int64 }
    CardPaymentSchedule:
      type: object
      required: [scheduleId, scheduledOn, status, attemptCount, nextRetryAt]
      properties:
        scheduleId: { type: string, format: uuid }
        scheduledOn: { type: string, format: date }
        status: { type: string, enum: [SCHEDULED, PROCESSING, COMPLETED, FAILED, CANCELLED] }
        attemptCount: { type: integer, minimum: 0 }
        nextRetryAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
    CardStatementPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CardStatementSummary" }
        nextCursor:
          oneOf:
            - { type: string }
            - { type: "null" }
    CreateCardStatementPrepaymentRequest:
      type: object
      required: [amountWon, expectedVersion]
      properties:
        amountWon: { type: integer, format: int64, minimum: 1 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    ApplyCardStatementPrepaymentRequest:
      allOf:
        - { $ref: "#/components/schemas/CreateCardStatementPrepaymentRequest" }
        - type: object
          required: [previewToken]
          properties:
            previewToken: { type: string, minLength: 1, maxLength: 64 }
    CardStatementPrepaymentPreview:
      type: object
      required: [previewToken, statementVersion, amountWon, appliedOn, currentRemainingAmountWon, afterRemainingAmountWon, settlementAsset, afterSettlementAssetBalanceWon]
      properties:
        previewToken: { type: string }
        statementVersion: { type: integer, format: int64, minimum: 0 }
        amountWon: { type: integer, format: int64, minimum: 1 }
        appliedOn: { type: string, format: date }
        currentRemainingAmountWon: { type: integer, format: int64, minimum: 1 }
        afterRemainingAmountWon: { type: integer, format: int64, minimum: 0 }
        settlementAsset: { $ref: "#/components/schemas/CardSettlementAsset" }
        afterSettlementAssetBalanceWon: { type: integer, format: int64 }
    CardStatementPaymentResult:
      type: object
      required: [statement, payment, settlementTransaction]
      properties:
        statement: { $ref: "#/components/schemas/CardStatementDetail" }
        payment: { $ref: "#/components/schemas/CardStatementPayment" }
        settlementTransaction: { $ref: "#/components/schemas/Transaction" }
...
    Problem:
      type: object
      required: [status, title, detail, errorCode, correlationId, timestamp, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errorCode: { type: string }
        correlationId: { type: string }
        timestamp: { type: string, format: date-time }
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: true
  responses:
    BadRequest:
      description: 입력값 또는 token 오류
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unauthorized:
      description: 인증 실패
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    Conflict:
      description: 현재 상태와 충돌하는 요청
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: 현재 사용자가 접근할 수 있는 리소스가 없음
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    PreconditionFailed:
      description: 편집 시작 후 다른 세션이 리소스를 수정함
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }